+++ .gitignore
@@ -3,4 +3,4 @@
 *.jl.mem
 docs/build/
 docs/site/
-benchmarks/data
+.results/
