+++ benchmark/benchmarks.jl
@@ -1,0 +1,35 @@
+using PkgBenchmark
+using LightGraphs
+
+
+testdatadir = joinpath(dirname(@__FILE__), "..","test","testdata")
+benchdatadir = joinpath(dirname(@__FILE__), "data")
+paramsfile = joinpath(benchdatadir, "params.jld")
+
+println("testdatadir = $testdatadir")
+println("paramsfile = $paramsfile")
+
+dg1fn = joinpath(testdatadir, "graph-5k-50k.jgz")
+
+DIGRAPHS = Dict{String, DiGraph}(
+    "complete100"   => CompleteDiGraph(100),
+    "5000-50000"    => LightGraphs.load(dg1fn)["graph-5000-50000"],
+    "path500"       => PathDiGraph(500)
+)
+
+GRAPHS = Dict{String, Graph}(
+    "complete100"   => CompleteGraph(100),
+    "tutte"         => smallgraph(:tutte),
+    "path500"       => PathGraph(500),
+    "5000-49947"    => Graph(DIGRAPHS["5000-50000"])
+)
+
+
+@benchgroup "LightGraphsBenchmarks" begin
+  include("core.jl")
+  include("edges.jl")
+  include("centrality.jl")
+  include("connectivity.jl")
+  include("max-flow.jl")
+  include("traversals.jl")
+end
